% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run}
\alias{run}
\title{Run Component Tools (Scina, scSorter, scType, scPred, Singler)}
\usage{
run(
  data_path,
  tools,
  markers = NULL,
  marker_names = NULL,
  ref_path = NULL,
  ref_labels_path = NULL
)
}
\arguments{
\item{data_path}{path to gene expression matrix stored as csv}

\item{tools}{TODO}

\item{markers}{marker genes for cell classification}

\item{marker_names}{TODO}

\item{ref_path}{path to refrence dataset for predictions}

\item{ref_labels_path}{path to refrence dataset labels}
}
\value{
compiled tool predictions
}
\description{
Run Component Tools (Scina, scSorter, scType, scPred, Singler)
}
