% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_r4scsharp}
\alias{run_r4scsharp}
\title{Run Script}
\usage{
run_r4scsharp(
  data_path,
  out_path,
  marker_path,
  ref_path,
  ref_label_path,
  tools = "scina,scsorter,sctype,singler,scpred",
  min_cells = 0,
  min_feats = 0
)
}
\arguments{
\item{data_path}{path to gene expression matrix stored as csv}

\item{out_path}{path to desired output save location}

\item{marker_path}{path to marker genes for cell classification}

\item{ref_path}{path to refrence dataset for predictions}

\item{ref_label_path}{path to refrence dataset labels}

\item{tools}{tools you would like to run (runs all tools by default)}

\item{min_cells}{filters data so cells with a sample size less then the
given amount are ignored (value is 0 by default)}

\item{min_feats}{filters data so features with a sample size less then the
given amount are ignored (value is 0 by default)}
}
\value{
compiled tool predictions
}
\description{
Run Script
}
