% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_scina}
\alias{run_scina}
\title{Runs Scina}
\usage{
run_scina(data, markers = NULL, ref = NULL)
}
\arguments{
\item{data}{Gene expression matrix}

\item{markers}{Marker genes for cell classification}

\item{ref}{Reference dataset for predictions.}
}
\value{
Returns Scina predictions as an R data frame
}
\description{
Runs Scina
}
