% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_scpred}
\alias{run_scpred}
\title{Run scPred}
\usage{
run_scpred(data, ref, ref_labels)
}
\arguments{
\item{data}{gene expression matrix}

\item{ref}{refrence dataset for predictions}

\item{ref_labels}{labels of refrence dataset}
}
\value{
scPred predictions
}
\description{
Run scPred
}
