% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_scsorter}
\alias{run_scsorter}
\title{Run scSorter}
\usage{
run_scsorter(data, markers = NULL, ref = NULL)
}
\arguments{
\item{data}{gene expression matrix}

\item{markers}{marker genes for cell classification}

\item{ref}{refrence dataset for predictions.}
}
\value{
scSorter predictions
}
\description{
Run scSorter
}
