% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_sctype}
\alias{run_sctype}
\title{Runs scType}
\usage{
run_sctype(data, markers = NULL, ref = NULL)
}
\arguments{
\item{data}{Gene expression matrix}

\item{markers}{Marker genes for cell classification}

\item{ref}{Refrence dataset for predictions.}
}
\value{
Returns scType predictions as an R data frame
}
\description{
Runs scType
}
