% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_singler}
\alias{run_singler}
\title{Run Singler}
\usage{
run_singler(data, ref, ref_labels)
}
\arguments{
\item{data}{gene expression matrix}

\item{ref}{refrence dataset for predictions}

\item{ref_labels}{labels of refrence dataset}
}
\value{
SingleR predictions
}
\description{
Run Singler
}
