% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{sctype_score}
\alias{sctype_score}
\title{sctype_score: calculate ScType scores and assign cell types}
\usage{
sctype_score(
  scRNAseqData,
  scaled = !0,
  gs,
  gs2 = NULL,
  gene_names_to_uppercase = !0,
  ...
)
}
\arguments{
\item{scRNAseqData}{\itemize{
\item input scRNA-seq matrix
(rownames - genes, column names - cells)
}}

\item{scaled}{\itemize{
\item indicates whether the matrix is scaled (TRUE by default)
}}

\item{gs}{\itemize{
\item list of gene sets positively expressed in the cell type
}}

\item{gs2}{\itemize{
\item list of gene sets that should not be expressed in the cell type
}}

\item{gene_names_to_uppercase}{\itemize{
\item if true, returns gene names in uppercase
}}

\item{...}{\itemize{
\item additional parameter input
(NULL if not applicable)
}}
}
\description{
sctype_score: calculate ScType scores and assign cell types
}
