% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_r4scsharp}
\alias{run_r4scsharp}
\title{R4scSHARP}
\usage{
run_r4scsharp(
  data_path,
  marker_path,
  ref_path,
  ref_label_path,
  out_path = NULL,
  tools = "scina,scsorter,sctype,singler,scpred",
  min_cells = 0,
  min_feats = 0
)
}
\arguments{
\item{data_path}{Path represented as a string to gene expression
matrix (or DGE matrix) stored as csv.}

\item{marker_path}{Path represented as a string to marker genes
for cell classification}

\item{ref_path}{Path represented as a string to reference dataset
for predictions}

\item{ref_label_path}{Path represented as a string to reference
dataset labels}

\item{out_path}{Path represented as a string to desired location for
the results to be saved. If no out_path parameter is given, no output
file will be generated.}

\item{tools}{Tools you would like to run (runs all tools by default).
Example Inputs:  "scina,scpred,singler", "sctype"}

\item{min_cells}{Filters data so cells with a sample size less then the
given amount are ignored (value is 0 by default)}

\item{min_feats}{Filters data so features with a sample size less then the
given amount are ignored (value is 0 by default)}
}
\value{
Returns the tool predictions in a single R data frame object
where the rows are cells, and the columns are tool prediction output values.
Save the returned data frame as a csv to a given output location.
}
\description{
This function is the primary function in the R4scSHARP
package. It calls another tool to run up to five cell classification
tools (SCINA, scSorter, scType, SingleR, and scPred) on an gene
expression matrix and saves the output file so scSHARP can run
predictions using that information.
}
