% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_scina}
\alias{run_scina}
\title{Runs scSHARP Prediction Toolsdocu
Run Scina}
\usage{
run_scina(data, markers = NULL, ref = NULL)
}
\arguments{
\item{data}{gene expression matrix}

\item{markers}{marker genes for cell classification}

\item{ref}{refrence dataset for predictions.}
}
\value{
scina predictions
}
\description{
Runs scSHARP Prediction Toolsdocu
Run Scina
}
