% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_scpred}
\alias{run_scpred}
\title{Runs scPred}
\usage{
run_scpred(data, ref, ref_labels)
}
\arguments{
\item{data}{Gene expression matrix}

\item{ref}{Reference dataset for predictions}

\item{ref_labels}{Labels of reference dataset}
}
\value{
Returns scPred predictions as an R data frame
}
\description{
Runs scPred
}
