% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_sctype}
\alias{run_sctype}
\title{Run scType}
\usage{
run_sctype(data, markers = NULL, ref = NULL)
}
\arguments{
\item{data}{gene expression matrix}

\item{markers}{marker genes for cell classification}

\item{ref}{refrence dataset for predictions.}
}
\value{
scType predictions
}
\description{
Run scType
}
