% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_singler}
\alias{run_singler}
\title{Runs Singler}
\usage{
run_singler(data, ref, ref_labels)
}
\arguments{
\item{data}{Gene expression matrix}

\item{ref}{Reference dataset for predictions}

\item{ref_labels}{Labels of reference dataset}
}
\value{
Returns SingleR predictions as an R data frame
}
\description{
Runs Singler
}
