% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_tools}
\alias{run_tools}
\title{Run Script}
\usage{
run_tools(
  data_path,
  out_path,
  marker_path,
  ref_path,
  ref_label_path,
  tools = "scina,scsorter,sctype,singler,scpred"
)
}
\arguments{
\item{data_path}{path to gene expression matrix stored as csv}

\item{out_path}{path to desired output save location}

\item{marker_path}{path to marker genes for cell classification}

\item{ref_path}{path to refrence dataset for predictions}

\item{ref_label_path}{path to refrence dataset labels}

\item{tools}{tool you would like to run}
}
\value{
compiled tool predictions
}
\description{
Run Script
}
