% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R4scSHARP.R
\name{run_tools}
\alias{run_tools}
\title{Run Component Tools (Scina, scSorter, scType, scPred, Singler)}
\usage{
run_tools(
  data_path,
  tools,
  min_cells,
  min_feats,
  markers = NULL,
  marker_names = NULL,
  ref_path = NULL,
  ref_labels_path = NULL
)
}
\arguments{
\item{data_path}{Path to gene expression matrix stored as csv}

\item{tools}{Tool you would like to run}

\item{min_cells}{Filters data so cells with a sample size less then the
given amount are ignored (value is 0 by default)}

\item{min_feats}{Filters data so features with a sample size less then the
given amount are ignored (value is 0 by default)}

\item{markers}{Marker genes for cell classification}

\item{marker_names}{Names of the marker genes}

\item{ref_path}{Path to refrence dataset for predictions}

\item{ref_labels_path}{Path to refrence dataset labels}
}
\value{
Returns the tool predictions in a single R data frame object
where the rows are cells, and the columns are tool prediction output values.
}
\description{
This function runs up to five cell classification
tools (SCINA, scSorter, scType, SingleR, and scPred) on an gene
expression matrix and returns the compiled predictions.
}
